logica lote

+++ model/LoteModel.java
@@ -1,5 +1,7 @@
 package com.myproyect.miproyect.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,9 +10,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class LoteModel {
-    private int idLote;//pk
+    private int idLote;// pk
     private String nombreLote;
-    private double longitudLote;
-    private double latitudLote;
+    private BigDecimal longitudLote;
+    private BigDecimal latitudLote;
     private String tipoDeCultivo;
 }
